backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
python-barcode==0.15.1
qrcode[pil]==8.0
Pillow==11.1.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
httpx==0.28.1
